refactor: rename packages and classes to follow new naming conventions
This commit renames several packages and classes to adhere to the new naming conventions. The changes include renaming the `com.thoughtworks.archguard` package to `org.archguard`, renaming the `JsonUtils` class to `JsonUtils` in the `org.archguard.json` package, renaming the `FeignClient` class to `FeignClient` in the `org.archguard.protocol.feigh` package, renaming the `HttpRequest` class to `HttpRequest` in the `org.archguard.protocol.http` package, renaming the `HttpRequestArg` class to `HttpRequestArg` in the `org.archguard.protocol.http` package, renaming the `FeignClientArg` class to `FeignClientArg` in the `org.archguard.protocol.feigh` package, and renaming the `QualityGateProfile` class to `QualityGateProfile` in the `com.thoughtworks.archguard.qualitygate.domain` package. These changes were made to improve code readability and maintainability.