refactor: update import statements and dependencies
This commit refactors the import statements and dependencies in multiple files. It updates the import statement in `DataClassRepositoryImpl.kt` to import `FieldVO` from the correct package. It also updates the import statement in `HttpRequestArg.kt` to remove the unused import of `RequestMethod`. Additionally, it updates the dependencies in `build.gradle.kts` files to use the latest version of `jackson-module-kotlin` and adds a new dependency for `kotlinstatistics`.