chore(documentation): add documentation for CircularDependenciesCount class
Added documentation for the CircularDependenciesCount class to provide information about its properties and usage.

The CircularDependenciesCount class is a data class that represents the count of circular dependencies in a system. It contains the following properties:
- systemId: The ID of the system.
- moduleCircularDependenciesCount: The count of circular dependencies at the module level.
- packageCircularDependenciesCount: The count of circular dependencies at the package level.
- classCircularDependenciesCount: The count of circular dependencies at the class level.
- methodCircularDependenciesCount: The count of circular dependencies at the method level.

This class is used to track and analyze circular dependencies in a software system.