feat(view/module): add buildArchitecture method to ArchitectureStyle
This commit adds a new method called `buildArchitecture` to the `ArchitectureStyle` interface. This method is implemented in all concrete classes that implement `ArchitectureStyle`. The purpose of this method is to build the architecture based on the provided `PotentialExecArch` object. Currently, the implementation of the method is empty in all concrete classes, but it can be overridden to perform specific actions based on the architecture style.