refactor(loader): skip loading analyser on Windows
This commit refactors the `AnalyserLoaderTest.kt` file to skip loading the analyser on Windows operating systems. This is necessary because the loading process is different on Windows and can cause issues. The code now checks the operating system name and returns early if it is Windows. This ensures that the tests are only executed on non-Windows systems.