feat(test): add DiffChangeControllerTest
This commit adds a new test class `DiffChangeControllerTest` to test the functionality of the `DiffChangeController`. The test class includes two test methods: `shouldReturnDiffChangesForHistoryInfluence` and `shouldExecuteDiffChangesForInfluenceByCommit`. These methods test the behavior of the controller when retrieving diff changes for history influence and executing diff changes for influence by commit, respectively. The test class is annotated with `@SpringBootTest` and `@ExtendWith(SpringExtension::class)`