feat(tests): add controller test classes
This commit adds three new test classes for controller testing: `ControllerTest.kt`, `PluginControllerTest.kt`, and `DruidStatControllerTest.kt`.

`ControllerTest.kt` is a new file that contains a test class for testing controllers. It is annotated with `@SpringBootTest` and `@ExtendWith(SpringExtension::class)` to enable Spring Boot testing. The test class sets up a `MockMvc` instance and contains some test methods.

`PluginControllerTest.kt` is an existing file that has been modified. The import statement for `WebMvcTest` has been removed, and the import statement for `status` has been added.

`DruidStatControllerTest.kt` is a new file that contains a test class for testing the `DruidStatController`. It is annotated with `@SpringBootTest`, `@ExtendWith(SpringExtension::class)`, and `@AutoConfigureMockMvc` to enable Spring Boot testing. The test class sets up a `MockMvc` instance and contains a test method that performs a GET request to the `/api/druid/stat` endpoint and expects a 200 status code.