feat(analyser): add Thrift IDL analyser and test
Add the Thrift IDL analyser to the build process and include a new test file for it. The analyser is now a part of the Continuous Deployment workflow and the Dockerfile. Additionally, the GoAnalyserTest has been refactored to remove redundant imports. A sample Thrift file is provided for testing purposes.