fix(scanner): ensure type safety for analysed code data structures #153
Previously, the code analysis process did not ensure the type safety of the data structures, potentially leading to runtime errors. This commit tightens the type definitions by explicitly casting the analysis results to `MutableList<CodeDataStruct>` and storing them in a variable of the same type. Additionally, the commit refines the logging message to generalise the error for IDL analysis.